+++ lib/scry_2/players/player.ex
@@ -5,6 +5,7 @@ defmodule Scry2.Players.Player do
   schema "players" do
     field :mtga_user_id, :string
     field :screen_name, :string
+    field :mtga_display_name, :string
     field :first_seen_at, :utc_datetime
 
     timestamps(type: :utc_datetime)
@@ -12,7 +13,7 @@ defmodule Scry2.Players.Player do
 
   def changeset(player, attrs) do
     player
-    |> cast(attrs, [:mtga_user_id, :screen_name, :first_seen_at])
+    |> cast(attrs, [:mtga_user_id, :screen_name, :mtga_display_name, :first_seen_at])
     |> validate_required([:mtga_user_id, :screen_name, :first_seen_at])
     |> unique_constraint(:mtga_user_id)
   end